modules/event.pl: track our own nick
As it seems, Net::IRC changes its idea of what our nick is when we *try*
to change nicks, not when the server acknowledges the change. This also
means it'll miss nick changes not done via ->nick (such as /ns regain or
;sl NICK ...). Therefore we'll have to react to incoming NICK messages
manually ... it's still not foolproof (because Net::IRC), but should fix
some minor bugs with the bot suffering an identity crisis and acting
strangely.

Notably, this should hopefully fix the issue with having to ;sync a
channel manually if it was ;join'ed while the bot didn't know its nick
properly. Maybe it'll fix some other weird crap, too.