BOOM GOES THE XML
This likely introduces twenty new bugs, but on the other hand it avoids
twenty thousand other issues (most of them relating to XML::Simple
doing the utterly wrong things all the time and configuration merely
mitigating the issues, rather than fixing the core problem)
so chances are it's more than worth it.

We check if there's shiny new JSON-based config in place already; if
not, we try to load XML config and convert it to the new format. The old
config files are left untouched in case we fucked up somehow.