Split 'pass' option into 'server_pass', 'account_pass'
Also remove the --pass command-line option for now and introduce a
backward compatibility check that warns when the old option is
encountered. This check is now placed with the mysql option checks in a
dedicated function.

---

So one day, I looked over the AntiSpamMeta console log and noticed
NickServ was echoing back part of the password to us in a "no such
username" error. This is due to us using the 'pass' option for both
nickserv and server password, so we sent PASS NickServPassword.
However, the password happened to have a ':' in it, so services
interpreted it as a user:password combination.