Make wordlist.txt, proxy.txt, nick_blacklist.txt optional
wordlist is used exclusively by garbagemeter which has been disabled for
quite a while. We'd probably be better off using a config-defined path
anyway so people can point the bot at /usr/share/dict/words or the local
equivalent.

The proxy list is most likely completely out of date. We have DNSBLs for
this anyway.

The nick blacklist is inflexible; nick blacklisting with an added reason
can already be done via the rules file.

Note that the functionality is not removed for now to avoid breaking
compatibility.