Move users to DB
Summary:
This will automatically take care of migrating the users from
users.json; you may delete that file.

Note that this removes htpasswd support. We now store (hashed) user
passwords in the database.

See T19 for rationale.

Test Plan: Run this on a testnet for a while, try to break it.

Reviewers: ilbelkyr, #antispammeta

Reviewed By: ilbelkyr, #antispammeta

Tags: #antispammeta, #database

Differential Revision: https://dev.antispammeta.net/D2